requirements.txt
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.17.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.18.0
six==1.16.0
requests==2.26.0
django-heroku
django-cors-headers
django-cleanup
django-storages
boto3
Pillow